# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-commerce API",
    "description": "API for user authentication, registration, and retrieving user details.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "description": "Allows users to register with a username, email, password, and role.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string", "example": "testuser" },
                  "email": { "type": "string", "example": "testuser@example.com" },
                  "password": { "type": "string", "example": "securepassword" },
                  "role": { "type": "string", "example": "customer", "default": "customer" }
                },
                "required": ["username", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "status_code": { "type": "integer" }
                  }
                },
                "example": {
                  "message": "Registration successful",
                  "status_code": 201
                }
              }
            }
          },
          "400": {
            "description": "Validation error or user already exists",
            "content": {
              "application/json": {
                "example": {
                  "message": "Email already registered",
                  "status_code": 400
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login a user",
        "description": "Authenticate a user with email and password, returning a JWT access token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "testuser@example.com" },
                  "password": { "type": "string", "example": "securepassword" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "status_code": { "type": "integer" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "access_token": { "type": "string" },
                        "role": { "type": "string" }
                      }
                    }
                  }
                },
                "example": {
                  "message": "Login successful",
                  "status_code": 200,
                  "data": {
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "role": "customer"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "example": {
                  "message": "Invalid credentials",
                  "status_code": 401
                }
              }
            }
          }
        }
      }
    },
    "/auth/user-details": {
      "get": {
        "summary": "Fetch user details",
        "description": "Retrieve details of the authenticated user. Requires a valid JWT token.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User details fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "status_code": { "type": "integer" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "name": { "type": "string" },
                        "email": { "type": "string" },
                        "role": { "type": "string" }
                      }
                    }
                  }
                },
                "example": {
                  "message": "User details fetched successfully",
                  "status_code": 200,
                  "data": {
                    "id": "cust-1732117897.301645",
                    "name": "testuser",
                    "email": "testuser@example.com",
                    "role": "customer"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "message": "Missing Authorization Header",
                  "status_code": 401
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "User not found",
                  "status_code": 404
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
     "/products/": {
      "post": {
        "summary": "Add a new product",
        "description": "Admin-only endpoint to add a new product to the catalog.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Product Name" },
                  "description": { "type": "string", "example": "Product Description" },
                  "price": { "type": "number", "example": 99.99 },
                  "stockQuantity": { "type": "integer", "example": 100 },
                  "category": { "type": "string", "example": "Electronics" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Product added successfully" },
          "400": { "description": "Validation error" }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "summary": "Get all products",
        "description": "Retrieve a paginated list of all products.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination."
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of products per page."
          }
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "example": "prod-12345" },
                          "name": { "type": "string", "example": "Product Name" },
                          "price": { "type": "number", "example": 99.99 },
                          "stockQuantity": { "type": "integer", "example": 50 },
                          "category": { "type": "string", "example": "Electronics" }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "currentPage": { "type": "integer", "example": 1 },
                        "totalPages": { "type": "integer", "example": 5 },
                        "totalItems": { "type": "integer", "example": 50 }
                      }
                    }
                  }
                },
                "example": {
                  "data": [
                    {
                      "id": "prod-12345",
                      "name": "Laptop",
                      "price": 999.99,
                      "stockQuantity": 10,
                      "category": "Electronics"
                    },
                    {
                      "id": "prod-67890",
                      "name": "Headphones",
                      "price": 199.99,
                      "stockQuantity": 25,
                      "category": "Accessories"
                    }
                  ],
                  "pagination": {
                    "currentPage": 1,
                    "totalPages": 2,
                    "totalItems": 20
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid pagination parameters"
          }
        }
      }
    },
    "/products/{product_id}": {
      "get": {
        "summary": "Get product details",
        "description": "Fetch details of a specific product by ID.",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Product details fetched successfully" },
          "404": { "description": "Product not found" }
        }
      },
       "put": {
        "summary": "Update a product",
        "description": "Admin-only endpoint to update the details of an existing product.",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the product to update."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Updated Product Name" },
                  "description": { "type": "string", "example": "Updated Product Description" },
                  "price": { "type": "number", "example": 89.99 },
                  "stockQuantity": { "type": "integer", "example": 150 },
                  "category": { "type": "string", "example": "Updated Category" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Product updated successfully" }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/json": {
                "example": {
                  "message": "Permission denied"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "example": {
                  "message": "Product not found"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "message": "Invalid input"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
    
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}